fixed deleteProduct to handle non-existent products

+++ server/src/controllers/product.js
@@ -74,8 +74,11 @@ const deleteProduct = async (request, response) => {
         try {
             const filter = {_id: request.params.productid, seller: seller}
             const productToDelete = await models.Product.find(filter)
-            await models.Product.deleteOne(filter)
-            return response.status(200).json({status: "successfully deleted product", productDeleted: productToDelete})
+            if(productToDelete) {
+                await models.Product.deleteOne(filter)
+                return response.status(200).json({status: "successfully deleted product", productDeleted: productToDelete})
+            }
+            return response.status(400).json({error: "product does not exist"})
         }
         catch {return response.status(401).json({error: "failed to delete product"})}
     }
